requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
backoff==2.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.4
discord.py==2.3.2
distro==1.9.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.352
langchain-community==0.0.17
langchain-core==0.1.18
langchain-openai==0.0.5
langfuse==2.2.2
langsmith==0.0.86
marshmallow==3.20.2
monotonic==1.6
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.11.1
packaging==23.2
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.1
redis==5.0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.0
spotipy==2.23.0
SQLAlchemy==2.0.25
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.0
wrapt==1.14.0
yarl==1.9.4
youtube_dl @ git+https://github.com/ytdl-org/youtube-dl.git@4416f82c809a81737d68875dcb201e366d58dabd
